--- a/tools/database/Cargo.toml
+++ b/tools/database/Cargo.toml
@@ -11,24 +11,29 @@ publish = false
 [dependencies]
 anyhow.workspace = true
 clap.workspace = true
+indicatif.workspace = true
+rand.workspace = true
 rayon.workspace = true
+rocksdb.workspace = true
 strum.workspace = true
 tempfile.workspace = true
 
 nearcore.workspace = true
-near-store.workspace = true
 near-chain-configs.workspace = true
-
+near-store.workspace = true
+near-primitives.workspace = true
 
 [features]
 nightly = [
   "nightly_protocol",
   "near-chain-configs/nightly",
+  "near-primitives/nightly",
   "near-store/nightly",
   "nearcore/nightly",
 ]
 nightly_protocol = [
   "near-chain-configs/nightly_protocol",
+  "near-primitives/nightly_protocol",
   "near-store/nightly_protocol",
   "nearcore/nightly_protocol",
 ]