--- a/tools/database/src/commands.rs
+++ b/tools/database/src/commands.rs
@@ -1,6 +1,7 @@
 use crate::adjust_database::ChangeDbKindCommand;
 use crate::analyse_data_size_distribution::AnalyseDataSizeDistributionCommand;
 use crate::make_snapshot::MakeSnapshotCommand;
+use crate::state_perf::StatePerfCommand;
 use clap::Parser;
 use std::path::PathBuf;
 
@@ -21,20 +22,17 @@ enum SubCommand {
 
     /// Make snapshot of the database
     MakeSnapshot(MakeSnapshotCommand),
+
+    /// Run performance test for State column reads.
+    /// Uses RocksDB data specified via --home argument.
+    StatePerf(StatePerfCommand),
 }
 
 impl DatabaseCommand {
     pub fn run(&self, home: &PathBuf) -> anyhow::Result<()> {
         match &self.subcmd {
             SubCommand::AnalyseDataSizeDistribution(cmd) => cmd.run(home),
-            SubCommand::ChangeDbKind(cmd) => {
-                let near_config = nearcore::config::load_config(
-                    &home,
-                    near_chain_configs::GenesisValidationMode::UnsafeFast,
-                )
-                .unwrap_or_else(|e| panic!("Error loading config: {:#}", e));
-                cmd.run(home, &near_config)
-            }
+            SubCommand::ChangeDbKind(cmd) => cmd.run(home),
             SubCommand::MakeSnapshot(cmd) => {
                 let near_config = nearcore::config::load_config(
                     &home,
@@ -43,6 +41,7 @@ impl DatabaseCommand {
                 .unwrap_or_else(|e| panic!("Error loading config: {:#}", e));
                 cmd.run(home, near_config.config.archive, &near_config.config.store)
             }
+            SubCommand::StatePerf(cmd) => cmd.run(home),
         }
     }
 }