--- a/tools/database/src/lib.rs
+++ b/tools/database/src/lib.rs
@@ -2,3 +2,5 @@ mod adjust_database;
 mod analyse_data_size_distribution;
 pub mod commands;
 mod make_snapshot;
+mod state_perf;
+mod utils;